PAGE_SIZE as usize

+++ kernel/src/memory/config.rs
@@ -1,6 +1,6 @@
 use x86_64::VirtAddr;
 
-pub const PAGE_SIZE: u64 = 4096;
+pub const PAGE_SIZE: usize = 4096;
 
 pub const KERNEL_START: VirtAddr = VirtAddr::new_truncate(0xFFFF_8000_0000_0000);
